requirements.txt
Flask==3.0.0
flask-cors==4.0.0
numpy==1.24.3
Pillow==10.2.0
Werkzeug==3.0.0
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.0+cpu
torchvision==0.17.0+cpu
